feat: 공통/파트 질문 데스크탑 & 모바일 뷰 통합
