+++ apps/ceos/src/pages/recruit/apply/Common.tsx
@@ -1,108 +1,40 @@
 import { Desktop, Flex, Mobile, Text, TextField } from 'packages/ui';
 import { RowLine, Section } from '.';
 import { RecruitApplyFormInterface } from './interface';
+import { CustomTextField, Explain, Question, QuestionFlex } from './style';
 
 interface CommonProps {
-  setValue: RecruitApplyFormInterface['setValue'];
+  register: RecruitApplyFormInterface['register'];
   questionList?: RecruitApplyFormInterface['questionList'];
 }
 
-const Common = ({ setValue, questionList }: CommonProps) => {
+const Common = ({ register, questionList }: CommonProps) => {
   return (
     <>
-      <Desktop>
-        <Section>
-          <Text webTypo="Heading3" paletteColor="Blue">
-            공통 질문
-          </Text>
-          {questionList?.commonQuestions.map((ques, idx) => (
-            <Flex
-              direction="column"
-              align="start"
-              webGap={12}
-              key={ques.questionId}
-            >
-              <Text webTypo="Label3">{`${idx + 1}. ${ques.question}`}</Text>
-              <TextField
-                width={856}
-                multiline={ques.multiline}
-                onChange={(e) =>
-                  setValue(`commonAnswers.${idx}.answer`, e.target.value)
-                }
-              />
-              <Flex direction="column" align="start">
-                {ques.questionDetail.map((detail, idx) =>
-                  detail.color === 'gray' ? (
-                    <Text
-                      webTypo="Body3"
-                      paletteColor="Gray5"
-                      key={`detail_${idx}`}
-                    >
-                      {detail.explaination}
-                    </Text>
-                  ) : (
-                    <Text
-                      webTypo="Body3"
-                      paletteColor="Blue"
-                      key={`detail_${idx}`}
-                    >
-                      {detail.explaination}
-                    </Text>
-                  ),
-                )}
-              </Flex>
+      <Section>
+        <Text webTypo="Heading3" paletteColor="Blue">
+          공통 질문
+        </Text>
+        {questionList?.commonQuestions.map((ques, idx) => (
+          <QuestionFlex key={ques.questionId}>
+            <Question>{`${idx + 1}. ${ques.question}`}</Question>
+            <CustomTextField
+              multiline={ques.multiline}
+              isMobileFull={true}
+              {...register(`commonAnswers.${idx}.answer`)}
+            />
+            <Flex direction="column" align="start">
+              {ques.questionDetail.map((detail, idx) =>
+                detail.color === 'gray' ? (
+                  <Explain key={`detail_${idx}`}>{detail.explaination}</Explain>
+                ) : (
+                  <Explain>{detail.explaination}</Explain>
+                ),
+              )}
             </Flex>
-          ))}
-        </Section>
-      </Desktop>
-      <Mobile>
-        <Section>
-          <Text mobileTypo="Heading2" paletteColor="Blue">
-            공통 질문
-          </Text>
-          {questionList?.commonQuestions.map((ques, idx) => (
-            <Flex
-              direction="column"
-              align="start"
-              mobileGap={14}
-              key={ques.questionId}
-            >
-              <Text mobileTypo="Heading4">{`${idx + 1}. ${
-                ques.question
-              }`}</Text>
-              <TextField
-                width={1000}
-                multiline={true}
-                onChange={(e) =>
-                  setValue(`commonAnswers.${idx}.answer`, e.target.value)
-                }
-              />
-              <Flex direction="column" align="start">
-                {ques.questionDetail.map((detail, idx) =>
-                  detail.color === 'gray' ? (
-                    <Text
-                      mobileTypo="Body2"
-                      paletteColor="Gray5"
-                      key={`detail_${idx}`}
-                    >
-                      {detail.explaination}
-                    </Text>
-                  ) : (
-                    <Text
-                      mobileTypo="Body2"
-                      paletteColor="Blue"
-                      key={`detail_${idx}`}
-                    >
-                      {detail.explaination}
-                    </Text>
-                  ),
-                )}
-              </Flex>
-            </Flex>
-          ))}
-        </Section>
-      </Mobile>
-
+          </QuestionFlex>
+        ))}
+      </Section>
       <RowLine />
     </>
   );
