+++ packages/ui/src/components/TextField/index.tsx
@@ -16,6 +16,7 @@ interface TextFieldProps
   width?: number;
   height?: number;
   multiline?: boolean;
+  isMobileFull?: boolean;
   isAdmin?: boolean;
   isSubTextField?: boolean;
   fontColor?: string;
@@ -46,6 +47,7 @@ export const TextField = forwardRef<
       height,
       multiline = false,
       isAdmin = false,
+      isMobileFull = false,
       isSubTextField = false,
       fontColor = theme.palette.Black,
       right,
